pyproject.toml
[project]
name = "mcr-ai-insights"
version = "0.1.0"
description = "Deterministic pipeline to normalize CMS MLR data into a modeling-ready issuer-state-market-year panel with optional inflation normalization."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "Proprietary" }
authors = [{ name = "Troy Alexander" }]
dependencies = [
  "pandas>=2.0",
  "numpy>=1.25",
  "pyarrow>=14.0",
  "python-dotenv>=1.0",
  "requests>=2.31",
  "tqdm>=4.66",
  "rich>=13.7",
  "typer>=0.12",
  "fredapi>=0.5.1",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
]

[project.scripts]
mcr-ai = "mcr_ai_insights.cli:app"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"